Frameworks Update
There is some weirdness when you try to switch between graph types in flot that I need to figure out. Added new buttons for the features I'm planning to have implemented by the end of Sprint 5. So far the options don't work right yet, need to fix the Pie Chart option and the Filter By Artist option for the demo tonight.